Move tests into separate dir, since they're integration tests
